pyproject.toml
[project]
name = "earningscall-framework"
version = "0.1.0"
description = ""
authors = [
    {name = "Alejandro Álvarez",email = "alejandro.alvrz.castro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "anytree (>=2.13.0,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "earningscall (>=1.2.1,<2.0.0)",
    "funasr (>=1.2.6,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "networkx (>=3.5,<4.0)",
    "numpy (>=1.24,<2.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "optuna (>=4.4.0,<5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pillow (==10.2.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "scikit-learn (>=1.3.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "torch (==2.2.2)",
    "transformers (==4.50.3)",
    "typer (>=0.16.0,<0.17.0)",
    "umap-learn (>=0.5.7,<0.6.0)",
    "torchaudio (==2.2.2)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "bertopic (>=0.17.4,<0.18.0)",
    "keybert (>=0.9.0,<0.10.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "tf-keras (>=2.20.1,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "statsmodels (>=0.14.6,<0.15.0)"
]

[tool.poetry.scripts]
earningscall-framework = "earningscall_framework.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
